+++ packages/template/src/routes.ts
@@ -1,19 +1,11 @@
 import { Routes } from '@angular/router';
 import { HomeComponent } from './pages/home/home.component';
-import { ChannelComponent } from './pages/channel/channel.component';
-import { DocViewerComponent, DocViewerHomeComponent } from './pages/doc-viewer/doc-viewer.component';
-import {
-    ComponentOverviewComponent,
-    ComponentApiComponent,
-    ComponentExamplesComponent,
-    ComponentEmptyComponent
-} from './pages/pages.module';
 
 const actualRoutes: Routes = [
     {
         path: '',
-        component: HomeComponent
-    }
+        component: HomeComponent,
+    },
 ];
 
 export const routes: Routes = [...actualRoutes];