+++ packages/template/src/services/global-context.ts
@@ -1,20 +1,22 @@
-import { Injectable, Inject, InjectionToken } from '@angular/core';
-import { DocgeniSiteConfig, NavigationItem, DocgeniMode, HomeDocMeta, CategoryItem } from '../interfaces/public-api';
+import { Injectable, Inject, InjectionToken, Signal, computed, WritableSignal, signal } from '@angular/core';
+import { DocgeniSiteConfig, NavigationItem, DocgeniMode, HomeDocMeta, CategoryItem, DocgeniTheme } from '../interfaces/public-api';
 import { HttpClient } from '@angular/common/http';
 import { languageCompare } from '../utils/language-compare';
 import { DOCUMENT, Location } from '@angular/common';
+
 export const CONFIG_TOKEN = new InjectionToken('DOC_SITE_CONFIG');
 
 export const DEFAULT_CONFIG: DocgeniSiteConfig = {
     title: 'Docgeni',
-    description: ''
+    description: '',
 };
 
 const DOCGENI_LOCALE_KEY = 'docgeni-locale';
 const DOCGENI_MODE_KEY = 'docgeni-mode';
+const DOCGENI_THEME_KEY = 'docgeni-theme';
 
 @Injectable({
-    providedIn: 'root'
+    providedIn: 'root',
 })
 export class GlobalContext {
     locale!: string;
@@ -29,6 +31,15 @@ export class GlobalContext {
 
     repo!: string;
 
+    theme: WritableSignal<DocgeniTheme> = signal(DocgeniTheme.light);
+
+    isDarkTheme: Signal<boolean> = computed(() => {
+        return (
+            (this.theme() === DocgeniTheme.system && window.matchMedia && window.matchMedia('(prefers-color-scheme: dark)').matches) ||
+            this.theme() === DocgeniTheme.dark
+        );
+    });
+
     get isDefaultLocale() {
         return this.locale === this.config.defaultLocale;
     }
@@ -37,7 +48,7 @@ export class GlobalContext {
         @Inject(CONFIG_TOKEN) public config: DocgeniSiteConfig,
         private http: HttpClient,
         @Inject(DOCUMENT) private document: any,
-        private location: Location
+        private location: Location,
     ) {
         this.setup();
     }
@@ -48,7 +59,7 @@ export class GlobalContext {
             return localeKeyFromUrl;
         } else {
             const cacheLocale = window.localStorage.getItem(DOCGENI_LOCALE_KEY) || window.navigator.language || '';
-            const locale = (this.config.locales || []).find(locale => {
+            const locale = (this.config.locales || []).find((locale) => {
                 return languageCompare(locale.key, cacheLocale);
             });
             if (locale) {
@@ -61,13 +72,14 @@ export class GlobalContext {
 
     private setup() {
         this.setLocale(this.getLocaleKey());
+        this.setTheme(this.getTheme());
 
         const cacheMode = window.localStorage.getItem(DOCGENI_MODE_KEY);
         if (cacheMode && ['lite', 'full'].includes(cacheMode)) {
             this.config.mode = cacheMode as DocgeniMode;
         }
 
-        document.body.classList.add(`dg-mode-${this.config.mode}`, `dg-theme-${this.config.theme}`);
+        document.body.classList.add(`dg-mode-${this.config.mode}`, `dg-navbar-theme-${this.config.theme}`);
         if (this.config.repoUrl) {
             const pattern = /https:\/\/github.com\/([^\/]*)\/([^\/]*)/.exec(this.config.repoUrl);
             if (pattern && pattern.length === 3) {
@@ -78,7 +90,7 @@ export class GlobalContext {
     }
 
     public getLocalKeyFromUrl() {
-        const localeFromUrl = (this.config.locales || []).find(locale => {
+        const localeFromUrl = (this.config.locales || []).find((locale) => {
             return this.location.path().startsWith(`/${locale.key}`);
         });
         return localeFromUrl && localeFromUrl.key;
@@ -89,6 +101,28 @@ export class GlobalContext {
         window.localStorage.setItem(DOCGENI_LOCALE_KEY, locale);
     }
 
+    private getTheme(): DocgeniTheme {
+        const cacheTheme = window.localStorage.getItem(DOCGENI_THEME_KEY) as DocgeniTheme;
+        if (cacheTheme && [DocgeniTheme.light, DocgeniTheme.dark, DocgeniTheme.system].includes(cacheTheme)) {
+            return cacheTheme;
+        } else {
+            return DocgeniTheme.light;
+        }
+    }
+
+    public setTheme(theme: DocgeniTheme) {
+        this.theme.set(theme);
+        window.localStorage.setItem(DOCGENI_THEME_KEY, theme);
+
+        if (this.isDarkTheme()) {
+            document.documentElement.setAttribute('theme', DocgeniTheme.dark);
+            document.documentElement.style.setProperty('color-scheme', 'dark');
+        } else {
+            document.documentElement.removeAttribute('theme');
+            document.documentElement.style.removeProperty('color-scheme');
+        }
+    }
+
     getNowTimestamp() {
         return new Date().getTime();
     }
@@ -96,9 +130,11 @@ export class GlobalContext {
     initialize() {
         return new Promise((resolve, reject) => {
             this.http
-                .get<{ navs: NavigationItem[]; docs: NavigationItem[]; homeMeta: HomeDocMeta }>(
-                    `assets/content/navigations-${this.locale}.json?t=${this.getNowTimestamp()}`
-                )
+                .get<{
+                    navs: NavigationItem[];
+                    docs: NavigationItem[];
+                    homeMeta: HomeDocMeta;
+                }>(`assets/content/navigations-${this.locale}.json?t=${this.getNowTimestamp()}`)
                 .subscribe({
                     next: (response: { navs: NavigationItem[]; docs: NavigationItem[]; homeMeta: HomeDocMeta }) => {
                         this.homeMeta = response.homeMeta;
@@ -106,9 +142,9 @@ export class GlobalContext {
                         this.docItems = this.sortDocItems(this.navs);
                         resolve(response);
                     },
-                    error: error => {
+                    error: (error) => {
                         reject(error);
-                    }
+                    },
                 });
         });
     }