+++ CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0-next.4](https://github.com/docgeni/docgeni/compare/v2.4.0-next.3...v2.4.0-next.4) (2024-09-19)
+
+
+### Features
+
+* banner support dark theme #TINFR-585 ([#561](https://github.com/docgeni/docgeni/issues/561)) ([ccf0bed](https://github.com/docgeni/docgeni/commit/ccf0bedda8c86b4974bff9a9432f190d57edb9d5)), closes [#TINFR-585](https://github.com/docgeni/docgeni/issues/TINFR-585)
+
+
+
+
+
+# [2.4.0-next.3](https://github.com/docgeni/docgeni/compare/v2.4.0-next.2...v2.4.0-next.3) (2024-09-03)
+
+
+### Features
+
+* support config and switch light,dark,system themes #TINFR-412 ([#559](https://github.com/docgeni/docgeni/issues/559)) ([6c57638](https://github.com/docgeni/docgeni/commit/6c5763862c7aac7ab730f07942557e591b72e46c)), closes [#TINFR-412](https://github.com/docgeni/docgeni/issues/TINFR-412)
+
+
+
+
+
 # [2.4.0-next.2](https://github.com/docgeni/docgeni/compare/v2.4.0-next.1...v2.4.0-next.2) (2024-08-19)
 
 **Note:** Version bump only for package docgeni