+++ packages/create-docgeni/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0-next.4](https://github.com/docgeni/docgeni/compare/v2.4.0-next.3...v2.4.0-next.4) (2024-09-19)
+
+**Note:** Version bump only for package create-docgeni
+
+
+
+
+
+# [2.4.0-next.3](https://github.com/docgeni/docgeni/compare/v2.4.0-next.2...v2.4.0-next.3) (2024-09-03)
+
+**Note:** Version bump only for package create-docgeni
+
+
+
+
+
 # [2.4.0-next.2](https://github.com/docgeni/docgeni/compare/v2.4.0-next.1...v2.4.0-next.2) (2024-08-19)
 
 **Note:** Version bump only for package create-docgeni